Preprocessing Environmental variables
The only issue with this code has to do with precipitation. I may need to edit this script since 50% of the precipitation data was missing, I imputed with the median which was equal to 0 inches of rainfall. This means that we are assuming the vast majority of days had 0 rainfall. Not sure what method of imputation people want to use to solve this.